requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
google-generativeai
Flask
fpdf2
reportlab
fastapi
